backend/requirements.txt
# ── Core ──
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
pydantic>=2.9.0,<3.0
pydantic-settings>=2.5.0,<3.0
python-multipart>=0.0.9
PyJWT>=2.9.0,<3.0
passlib[bcrypt]>=1.7.4,<2.0
sse-starlette>=2.1.0,<3.0
aiofiles>=24.1.0

# ── Database ──
sqlalchemy[asyncio]>=2.0.32,<3.0
asyncpg>=0.29.0,<1.0
alembic>=1.13.0,<2.0
pgvector>=0.3.0,<1.0

# ── Vector Store ──
qdrant-client>=1.11.0,<1.13.0

# ── Graph Store ──
neo4j>=5.24.0,<6.0

# ── Cache & Queue ──
redis>=5.0.0,<6.0
celery[redis]>=5.4.0,<6.0

# ── LLM Providers ──
openai>=1.44.0,<2.0
anthropic>=0.34.0,<1.0
litellm>=1.44.0,<2.0

# ── Embeddings ──
sentence-transformers>=3.0.0,<4.0
fastembed>=0.3.0,<1.0

# ── Document Processing ──
langchain-text-splitters>=0.2.0,<1.0
unstructured[pdf,docx]>=0.15.0,<1.0
pypdf>=4.3.0,<5.0
python-docx>=1.1.0,<2.0
python-magic>=0.4.27,<1.0

# ── RAG Components ──
ragas>=0.1.18,<1.0
rank-bm25>=0.2.2,<1.0

# ── Observability (pinned: all otel packages must match) ──
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-semantic-conventions==0.48b0
prometheus-fastapi-instrumentator>=7.0.0,<8.0
structlog>=24.4.0

# ── Security ──
cryptography>=43.0.0
python-dotenv>=1.0.0,<2.0

# ── Testing ──
httpx>=0.27.0,<1.0
pytest>=8.3.0,<9.0
pytest-asyncio>=0.24.0,<1.0
pytest-cov>=5.0.0,<6.0
